lobby reconnect fix

+++ src/app/routers/tvt/map_selector.py
@@ -30,7 +30,8 @@ async def websocket_lobby_selector(websocket: WebSocket, tournament_id: int = Qu
     try:
         await lobby_selector_lifecycle(websocket, user, tournament_id)
     except WebSocketDisconnect:
-        tvt_service.remove_from_wait_room(data.email, tournament_id)
+        if TournamentInternalStateManager.get_state(tournament_id) == TournamentInternalStateManager.State.CONNECTING:
+            tvt_service.remove_from_wait_room(data.email, tournament_id)
         logger.info(f'[lobby selector] disconnected {user.team_name}')
